updated packages and manually merged pull #45
