--- src/main/groovy/org/owasp/dependencycheck/extension/PurgeDataExtension.groovy
@@ -1,29 +1,0 @@
-/*
- * This file is part of dependency-check-gradle.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Copyright (c) 2015 Jeremy Long. All Rights Reserved.
- */
-
-package org.owasp.dependencycheck.extension
-
-/**
- * The data configuration extension. Any value not configured will use the dependency-check-core defaults.
- */
-class PurgeDataExtension {
-    /**
-     * The directory to store the H2 database that contains the cache of the NVD CVE data.
-     */
-    String directory="[JAR]/../../dependency-check-data/3.0"
-}
