--- src/main/groovy/org/owasp/dependencycheck/extension/UpdateExtension.groovy
@@ -1,33 +1,0 @@
-/*
- * This file is part of dependency-check-gradle.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Copyright (c) 2015 Wei Ma. All Rights Reserved.
- */
-
-package org.owasp.dependencycheck.extension
-
-class UpdateExtension extends PurgeExtension {
-    ProxyExtension proxyExtension
-    CveExtension cveExtension
-    DataExtension dataExtension
-    /**
-     * Set to false if the proxy does not support HEAD requests. The default is true.
-     */
-    Boolean quickQueryTimestamp
-    /**
-     * The number of hours to wait before checking for additional updates from the NVD.
-     */
-    Integer cveValidForHours
-}
