+++ src/main/groovy/org/owasp/dependencycheck/gradle/extension/ProxyExtension.groovy
@@ -16,9 +16,14 @@
  * Copyright (c) 2015 Wei Ma. All Rights Reserved.
  */
 
-package org.owasp.dependencycheck.extension
-
-class PurgeExtension {
-    PurgeDataExtension dataExtension
-    Boolean failOnError = true
+package org.owasp.dependencycheck.gradle.extension
+/**
+ * TODO - this should not be needed, instead rely on the configured HTTP or HTTPS proxies
+ * https://docs.gradle.org/current/userguide/build_environment.html
+ */
+class ProxyExtension {
+    String server
+    Integer port
+    String username
+    String password
 }
