+++ src/main/groovy/org/owasp/dependencycheck/gradle/tasks/Update.groovy
@@ -16,7 +16,7 @@
  * Copyright (c) 2015 Jeremy Long. All Rights Reserved.
  */
 
-package org.owasp.dependencycheck.tasks
+package org.owasp.dependencycheck.gradle.tasks
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
@@ -52,7 +52,7 @@ import static org.owasp.dependencycheck.utils.Settings.KEYS.DB_PASSWORD
  */
 class Update extends DefaultTask {
 
-    @Internal def config = project.dependencyCheckUpdate
+    @Internal def config = project.dependencyCheck
 
     /**
      * Initializes the update task.
